fix(develop): fix observer

+++ app/src/main/java/org/example/services/MenuService.java
@@ -3,6 +3,7 @@ package org.example.services;
 import org.example.models.Menu;
 import org.example.models.Plate;
 import org.example.models.Restaurant;
+import org.example.observer.TheObserver;
 import org.example.repositories.MenuRepository;
 import org.example.repositories.RestaurantRepository;
 
@@ -20,12 +21,12 @@ public class MenuService {
     public void addRestaurantPlate(String restaurantName, String plateName, Double price) {
         Restaurant restaurant = restaurantRepository.getRestaurant(restaurantName);
         Plate plate = new Plate(plateName, price);
+        plate.addObserver(new TheObserver());
 
         if (restaurant == null) {
             System.out.println("Restaurante no encontrado. Creando nuevo restaurante...");
             restaurant = new Restaurant(restaurantName);
             restaurantRepository.addRestaurant(restaurant);
-
             Menu menu = new Menu(restaurant);
             menuRepository.addMenu(menu);
         }
@@ -41,11 +42,19 @@ public class MenuService {
     public void deleteRestaurantPlate(String restaurantName, String plateName) {
         Restaurant restaurant = restaurantRepository.getRestaurant(restaurantName);
         if (restaurant != null) {
-            menuRepository.deletePlateFromMenu(restaurant, plateName);
+            Plate plate = restaurant.getMenu().getPlateByName(plateName);
+            if (plate != null) {
+                plate.removeAllObservers();
+
+                menuRepository.deletePlateFromMenu(restaurant, plateName);
+            } else {
+                System.out.println("Plato no encontrado en el menu del restaurante.");
+            }
         } else {
             System.out.println("Restaurante no encontrado.");
         }
     }
+
     public boolean editRestaurantPlate(String restaurantName, String plateName, String newPlateName, Double newPrice) {
         Restaurant restaurant = restaurantRepository.getRestaurant(restaurantName);
         if (restaurant == null) {
@@ -58,9 +67,9 @@ public class MenuService {
     public void viewPlatesInRestaurant(String restaurantName) {
         Set<Plate> plates = menuRepository.getPlatesByRestaurantName(restaurantName);
         if (plates == null || plates.isEmpty()) {
-            System.out.println("No hay platos en el menú del restaurante " + restaurantName);
+            System.out.println("No hay platos en el menu del restaurante ");
         } else {
-            System.out.println("Platos en el menú de " + restaurantName + ":");
+            System.out.println("Platos en el menu de " + restaurantName + ":");
             plates.forEach(plate -> System.out.println(plate.getPlateName() + " - $" + plate.getPrice()));
         }
     }
